medical-stt-postprocess/requirements.txt
torch
transformers
jamo
python-Levenshtein
